Reorganize stop details page folders
One folder for each section, with additional subfolders if needed,
and one folder for shared layout components.
