+++ ui/src/components/stop-registry/stops/stop-details/layout/HorizontalSeparator.tsx
