+++ ui/src/components/stop-registry/stops/stop-details/layout/LabeledDetail.tsx
