+++ ui/src/components/stop-registry/stops/stop-details/location-details/LocationDetailsSection.tsx
@@ -1,6 +1,6 @@
 import { useTranslation } from 'react-i18next';
-import { StopWithDetails, useToggle } from '../../../../hooks';
-import { ExpandableInfoContainer } from './ExpandableInfoContainer';
+import { StopWithDetails, useToggle } from '../../../../../hooks';
+import { ExpandableInfoContainer } from '../layout';
 import { LocationDetailsViewCard } from './LocationDetailsViewCard';
 
 interface Props {
