+++ ui/src/components/stop-registry/stops/stop-details/location-details/index.ts
@@ -1,0 +1,2 @@
+export * from './LocationDetailsSection';
+export * from './LocationDetailsViewCard';
